rename actor.analytics package to actor.events
