+++ game-service-actor/src/test/scala/com/andy327/actor/battleship/BattleshipActorSpec.scala
@@ -7,8 +7,8 @@ import org.apache.pekko.actor.typed.ActorRef
 import org.scalatest.matchers.should.Matchers
 import org.scalatest.wordspec.AnyWordSpecLike
 
-import com.andy327.actor.analytics.NoOpAnalyticsPublisher
 import com.andy327.actor.core.{GameManager, PlayerActor, PlayerEvent, TurnBasedGameActor}
+import com.andy327.actor.events.NoOpAnalyticsPublisher
 import com.andy327.actor.game.{BattleshipState, GameState}
 import com.andy327.actor.persistence.PersistenceProtocol
 import com.andy327.model.battleship.{Coord, Fire}
