+++ game-service-server/src/main/scala/com/andy327/server/analytics/GameMetrics.scala
@@ -2,8 +2,8 @@ package com.andy327.server.analytics
 
 import io.prometheus.client.{CollectorRegistry, Counter, Histogram}
 
-import com.andy327.actor.analytics.GameAnalyticsEvent
-import com.andy327.actor.analytics.GameAnalyticsEvent._
+import com.andy327.actor.events.GameAnalyticsEvent
+import com.andy327.actor.events.GameAnalyticsEvent._
 import com.andy327.model.core.GameType
 
 /** Prometheus metrics for game analytics. Registers a small, low-cardinality set of collectors on `registry` and
