+++ game-service-server/src/main/scala/com/andy327/server/analytics/RedisAnalyticsPublisher.scala
@@ -5,7 +5,7 @@ import cats.effect.unsafe.IORuntime
 
 import io.circe.syntax._
 
-import com.andy327.actor.analytics.{AnalyticsPublisher, GameAnalyticsEvent}
+import com.andy327.actor.events.{AnalyticsPublisher, GameAnalyticsEvent}
 import com.andy327.server.analytics.GameAnalyticsCodecs.encoder
 
 object RedisAnalyticsPublisher {
