+++ game-service-server/src/test/scala/com/andy327/server/analytics/GameMetricsSpec.scala
@@ -6,7 +6,7 @@ import io.prometheus.client.CollectorRegistry
 import org.scalatest.matchers.should.Matchers
 import org.scalatest.wordspec.AnyWordSpec
 
-import com.andy327.actor.analytics.GameAnalyticsEvent._
+import com.andy327.actor.events.GameAnalyticsEvent._
 import com.andy327.model.core.GameType
 
 class GameMetricsSpec extends AnyWordSpec with Matchers {
