+++ game-service-server/src/test/scala/com/andy327/server/http/routes/MetricsRoutesSpec.scala
@@ -9,7 +9,7 @@ import org.apache.pekko.http.scaladsl.testkit.ScalatestRouteTest
 import org.scalatest.matchers.should.Matchers
 import org.scalatest.wordspec.AnyWordSpec
 
-import com.andy327.actor.analytics.GameAnalyticsEvent
+import com.andy327.actor.events.GameAnalyticsEvent
 import com.andy327.model.core.GameType
 import com.andy327.server.analytics.GameMetrics
